LPS-61979 Move common logic to base class
